[SPARK-44325][SQL] Use PartitionEvaluator API in SortMergeJoinExec
### What changes were proposed in this pull request?

SQL operator `SortMergeJoinExec` updated to use the `PartitionEvaluator` API to do execution.

### Why are the changes needed?

To avoid the use of lambda during distributed execution.
Ref: SPARK-43061 for more details.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Updated 1 test case, once all the SQL operators are migrated, the flag `spark.sql.execution.useTaskEvaluator` will be enabled by default to avoid running the tests with and without this TaskEvaluator

Closes #41884 from vinodkc/br_refactorSortMergeJoinEvaluatorFactory.

Authored-by: Vinod KC <vinod.kc.in@gmail.com>
